Get provisioner working!

+++ lib/setup.sh
@@ -1,5 +1,5 @@
 # Install xcode command line tools; required for homebrew, xcode, and maybe some other applications
-xcode-select --install
+xcode-select --install || echo "An error occurred installing xcode tools; proceeding with other steps."
 
 source "$ROCK_DIR"/lib/homebrew.sh
 
@@ -10,14 +10,20 @@ mkcert -install
 
 # Link tableplug config
 
-# Merge vscode settings with user's
-VSCODE_SETTINGS=/Library/"Application Support"/Code/User/settings.json
-jq -s '.[0] * .[1]' "$ROCK_DIR"/config/.vscode/settings.json "$VSCODE_SETTINGS" >/tmp/vscode_settings
-mv /tmp/vscode_settings "$VSCODE_SETTINGS"
+# Merge vscode settings with user's, if they exist
+VSCODE_SETTINGS="$ROCK_DIR"/config/.vscode/settings.json
+VSCODE_USER_SETTINGS=/Library/"Application Support"/Code/User/settings.json
+if [[ -f $VSCODE_USER_SETTINGS ]]; then
+    jq -s '.[0] * .[1]' "$VSCODE_SETTINGS" "$VSCODE_USER_SETTINGS" >/tmp/vscode_settings
+    sudo cp /tmp/vscode_settings "$VSCODE_USER_SETTINGS"
+else
+    sudo mkdir -p "$(dirname "$VSCODE_USER_SETTINGS")"
+    sudo cp "$VSCODE_SETTINGS" "$VSCODE_USER_SETTINGS"
+fi
 # Install vscode extensions
 while IFS= read -r LINE; do
     code --install-extension "$LINE"
-done <"ROCK_DIR"/.vscode/extensions.txt
+done <"$ROCK_DIR"/config/.vscode/extensions.txt
 
 # Setup oh-my-zsh & its plugins
 source "$ROCK_DIR"/lib/oh-my-zsh.sh
@@ -25,19 +31,10 @@ source "$ROCK_DIR"/lib/oh-my-zsh.sh
 # Setup vim & its plugins
 source "$ROCK_DIR"/lib/vim-plug.sh
 
-# Link vim config
-INSTALL_VIMRC=$(get_option vimrc)
-if $INSTALL_VIMRC; then
-    sed -i '' "1s-^-let \$VIM_DIR='$VIM_DIR' \" Managed by rock\n-g" ~/.vimrc
-    sed -i '' "2s-^-source $ROCK_DIR/config/vimrc \" Managed by rock\n-g" ~/.vimrc
-    sed -i '' "3s-^-\" Include your own configurations below this line \" Managed by rock\n-g" ~/.vimrc
-fi
-
-echo "Linking vim config..."
-ln -f "$ROCK_DIR"/config/vimrc /etc/vim/vimrc.local
-
 echo "Ensuring updatedb is on PATH"
-ln -s /usr/libexec/locate.updatedb /usr/local/bin/updatedb
+if [[ ! -f /usr/local/bin/updatedb ]]; then
+    ln -s /usr/libexec/locate.updatedb /usr/local/bin/updatedb
+fi
 
 echo "Ensuring rock is on PATH..."
 if [[ ! -f /usr/local/bin/rock ]]; then
